Add command line tools for GA Reporting API
This commit adds command line tools for invoking the GA Reporting API
(v4) and parsing its responses. A requirements.txt file recording the
full python library dependencies list is also included.

The invoker (ga-reporting.py) accepts a json file containing the report
request body, and produces a JSON to stdout containing the API response.
Authentication is handled through the standard google_auth.py handler.

The parser (ga-parse.jq) parses the reporting API response into TSV
output on stdout. Not all report response features are supported yet.
Reports containing pivots or multiple date ranges won't parse
correctly. The current metric data type parsing is naive and assumes
that the requested metrics should be of integer type. The parser doesn't
check for sampling, or whether the data is golden (set in stone).

Currently, the parser treats multiple reports in a single batch as an
outer join. The dimensions are assumed to be the same, and all of the
metrics are merged into a single output row. Additional batch handling
options are planned for future updates.