Create library of common GA API parsing functions
This commit establishes a standard library of useful jq parsing
functions for handling responses from the GA reporting API. This
includes conversions between Tabular <--> Object Oriented data formats,
rowwise joins, and column mappings.

This commit also establishes a directory for programs & data files
associated with testing. The ga-parse.jq file was renamed & moved to
this file to emphasize its role as a testing utility. In addition, it
was refactored to utilize the new standard library of parsing functions.