Basic GA reporting response parser + cleanup
This commit adds a default parser that can be used for parsing simple
GA Reporting API responses involving a single report.

In addition, some changes were made to tidy up the project. The report
request builder was moved to the utilities folder to clean up the main
directory. The requirements.txt file was updated to allow upgrades to
newer versions of libraries.

The Google authentication script was updated to comment out test cases
invoking the Google Ads API. This is intended to be a temporary change
until a new Google Ads developer token can be setup on the new account;
until then, all Google Ads related tests would fail.