Add batch mutate feature GA Management CLI
This commit introduces a backwards-incompatible interface change to
support multiple mutate operations within a single command invocation.
In addition, the output format was changed to JSONL (from JSON) to
support logging multiple API responses from a single command invocation.

The argument list for patch / update operations was changed. Previously,
users had to specify the entity ID to operate on in both the command
line and on the actual object body itself, to match the API that both
requires the entity ID as part of the request URL, and as part of the
object body.

The CLI will now read the entity ID from the object body,
removing the need to collect a duplicate entity ID from the command
line arguments. The CLI arguments on patch / update operations therefore
no longer require an entity ID to be specified, and specifying one
produces an error.

With the interface change, I have added the ability to pass multiple
object bodies for mutation in a single command invocation. If passing in
multiple object bodies, the input must be a single JSONL document, 1
object body per line. This change is backwards compatible - passing a
pretty-printed JSON file will still work as long as there's only 1 top
level object in the input file.

The output format has changed from JSON to JSONL, to support logging
multiple API responses from batch mutate requests in a
backwards-compatible manner.

This change also fixes a bug in the default handling of the "body"
argument, which was updated to "stdin" (from "stdout") in both the
ga-management.py and the tagmanager.py CLIs.