Move discovery docs & API request function
This commit introduces 2 primary changes:
* All discovery document related parsers and documents have been added
  to a dedicated directory
* The shared "send_request" functionality between all the API parsers
  has been moved to the shared "google_auth.py" library

The new "discovery" directory groups all of the Google Discovery related
documents and parsers under a single directory. The parsers directory
contains the parsers for the discovery documents, which is the only part
of the discovery documents directory actually committed. The documents
themselves will eventually be added to an installation makefile that
downloads / parses the needed documents on install. All 3 API invocation
scripts were updated to point to the new discovery document location.

The "send_request" function containing the exponential backoff / HTTP
error handling was shared between multiple scripts, so it was moved to
the shared "google_auth.py" document. This may be moved later to a
separate library file if more such utility functions are added
over time.