Version upgrade & minor bugfixes
This commit migrates the Google APIs codebase to Python 3.11. The
library requirements have been updated accordingly.

In addition, a few minor changes were made to the authentication &
tagmanager libraries. The tagmanager library was updated to explicitly
call for a string conversion of a Path variable, which was previously
implicitly occurring. This resolves a false positive in the mypy type
checker. The authentication library had a few comments added, and a
method stub was updated to remove a type checking error on an
currently unused method.