Universal query builder initial version
This commit includes an initial version of a parser for a unified query
specification format. Currently, this query format allows users to
specify queries for a modest subset of the Universal Analytics Reporting
API. This commit includes an additional module defining types & an
expression parser that are used in the main uar.py module.