Refactor UA reporting API CLI bindings
This commit is primarily concerned with standardizing code styles in the
UA reporting API CLI bindings script. In addition to this category of
refactors, a new CLI option was added to treat API query input as either
a pre-formed API request, or to use the new universal query format.
Full support for the new query format is to be added in a later update.