Refactor UA CLI & Expression Parser
This commit includes refactors of 2 major components. The refactors in
ga-reporting compartmentalize the control flow for handling raw API
query inputs, in preparation for adding support for new query definition
formats.

The refactors in uar_types.py roll the construction process for the
Expression parser into a function that returns the completed expression
parser. This eliminates a large number of global variables generated
during the construction process.