Add DateRange iteration feature
This commit adds a handful of utility methods to the DateRange class.
DateRanges now have a length property that returns the number of days in
the date range. Users can ask whether a date object (or ISO-formatted
date string) is in the date range. And users can query for the n'th date
in the date range, or a sub-interval of the date range. Slices passed to
getitem are expected to have a step size of 1 -- other step sizes are
not supported.